+++ src/example/ksi_verify_pub.c
@@ -211,7 +211,7 @@ int main(int argc, char **argv) {
 		}
 		/* Calculate the hash of the document. */
 		while (!feof(in)) {
-			buf_len = (unsigned)fread(buf, 1, sizeof(buf), in);
+			buf_len = fread(buf, 1, sizeof(buf), in);
 			res = KSI_DataHasher_add(hsr, buf, buf_len);
 			if (res != KSI_OK) {
 				fprintf(stderr, "Unable hash the document.\n");
